create custom hook - useGetFolder hook을 제작했습니다.
